+++ src/main/java/com/malbano/rural/model/entity/DadosEntity.java
@@ -34,7 +34,7 @@ public class DadosEntity {
     @JsonProperty("QtdCusteio")
     private Integer qtdCusteio;
     @JsonProperty("VlCusteio")
-    private BigDecimal vlCusteio;
+    private Double vlCusteio;
     @JsonProperty("Atividade")
     private String atividade;
     private String cdModalidade;
